usb: host: ohci-at91: add support for SMC to access SFR
When Linux is running under OP-TEE, the SFR is set a secured and thus
the AT91_OHCIICR_USB_SUSPEND register isn't accessible. Add a SMC to
do the appropriate call to suspend the controller.
The SMC id is fetched from the device-tree property
microchip,suspend-smc-id. if present, then the syscon regmap is not
used to enter suspend.

Signed-off-by: Clément Léger <clement.leger@bootlin.com>