ARM: dts: at91: sama5d27_wlsom1_ek: Add no-1-8-v property to sdmmc0 node
Add no-1-8-v property to sdmmc0 node to keep VDDSDMMC power rail at
3.3V. This property will stop the LDO regulator from switching to 1.8V
when the MMC core detects an UHS SD Card. VDDSDMMC power rail is used
by all the SDMMC interface pins in GPIO mode (PA0 - PA13).

On this board, PA10 is used as GPIO to enable the power switch controlling
USB Vbus for the USB Host. The change is needed to fix the PA10 voltage
level to 3.3V instead of 1.8V.

Fixes: 5d4c3cfb63fe ("ARM: dts: at91: sama5d27_wlsom1: add SAMA5D27 wlsom1 and wlsom1-ek")
Suggested-by: Mihai Sain <mihai.sain@microchip.com>
Signed-off-by: Cristian Birsan <cristian.birsan@microchip.com>