ARM: at91: pm: fix MCK1..9 save/restore for sama7g5/sama7d65 SOC
The PMC_MCR_CMD bit needs to be set for updating the value of PMC_MCR
in write mode.
 - Set PMC_MCR_CMD to '1' in the save macro.
 - Use the correct variable to write to PMC_MCR in the restore macro.
Switch MCK1..4 to MD_SLCK(divided by 64) according to Table "Low-Power
Mode Summary" in datasheet. As the OSC_STBY is controlled by the LPM pin
on sama7g54-ek and sama7d65-curiosity boards enabling LPM will hang
the processor when MCK is switched to MAINCK.

Signed-off-by: Tony Han <tony.han@microchip.com>
[romain.sioen@microchip.com: adapted to sama7d65 code with 9 mck]
Signed-off-by: Romain Sioen <romain.sioen@microchip.com>
Tested-by: Romain Sioen <romain.sioen@microchip.com> # on sama7g5 and sama7d65