net: dsa: microchip: add suspend/resume support to KSZ DSA driver
The KSZ DSA driver starts a timer to read MIB counters periodically to
avoid count overrun.  When the system is suspended the SPI register
write returns an error because the SPI system now displays an error
message when it is suspended.  This patch stops the timer when the driver
is suspended and restarts the timer when resumed.

Signed-off-by: Tristram Ha <tristram.ha@microchip.com>
Tested-by: Hari Prasath Gujulan Elango <Hari.PrasathGE@microchip.com>